// src/app/+admin/services/interfaces/IRolePageAssociate.ts
export interface IRolePageAssociate {
    rolePageAssocId?: number | null | undefined;

    applicationId?: number | null | undefined;
    application?: any | null | undefined;

    roleId?: number | null | undefined;
    role?: any | null | undefined;

    pageId?: number | null | undefined;

    isActive?: boolean | null | undefined;
    userId?: number | null | undefined;
    ipAddress?: string | null | undefined;
}